test : Post Author 수정 - WithMockCustomUser 사용시 db에 저장된 user 정보와 다를 수 있어서 user 새로 생성 후 authentication 수정
- post 저장전 author를 저장하기 위해 createEntity 에 user 파라미터 추가ㅣ

+++ server/src/main/java/com/choo/blog/domain/posts/service/PostService.java
@@ -3,16 +3,24 @@ package com.choo.blog.domain.posts.service;
 import com.choo.blog.domain.posts.Post;
 import com.choo.blog.domain.posts.dto.PostRequestData;
 import com.choo.blog.domain.posts.repository.PostRepository;
+import com.choo.blog.domain.users.User;
+import com.choo.blog.domain.users.repository.UserRepository;
+import com.choo.blog.exceptions.ForbiddenPostException;
 import com.choo.blog.exceptions.PostNotFoundException;
+import com.choo.blog.exceptions.UserNotFoundException;
+import com.choo.blog.security.UserAuthentication;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 @RequiredArgsConstructor
 @Service
 public class PostService {
     private final PostRepository postRepository;
+
+    private final UserRepository userRepository;
 
     public Page<Post> getPosts(Pageable pageable){
         return postRepository.findAll(pageable);
@@ -23,18 +31,36 @@ public class PostService {
     }
 
     public Post save(PostRequestData saveData){
-        return postRepository.save(saveData.createEntity());
+        User author = getLoginUser();
+
+        return postRepository.save(saveData.createEntity(author));
     }
 
     public Post update(Long id, PostRequestData updateData){
         Post posts = postRepository.findById(id).orElseThrow(() -> new PostNotFoundException(id));
+        User user = getLoginUser();
+
+        if(!posts.getAuthor().getId().equals(user.getId())){
+            throw new ForbiddenPostException(posts.getId());
+        }
         posts.update(updateData);
 
         return posts;
     }
 
     public void delete(Long id){
-        Post post = getPost(id);
+        Post post = getPost(id);User user = getLoginUser();
+
+        if(!post.getAuthor().getId().equals(user.getId())){
+            throw new ForbiddenPostException(post.getId());
+        }
+
         postRepository.delete(post);
+    }
+
+    private User getLoginUser(){
+        UserAuthentication authentication = (UserAuthentication) SecurityContextHolder.getContext().getAuthentication();
+        return userRepository.findById(authentication.getUserId())
+                .orElseThrow(() -> new UserNotFoundException(authentication.getUserId()));
     }
 }
